pyproject.toml
[tool.poetry]
name = "langgraph-course"
version = "0.1.0"
description = ""
authors = ["Eden Marco <emarco177@gmail.com>"]
readme = "README.md"
packages = [{ include = "graph" }]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.3"
chromadb = "^0.5.0"
tavily-python = "^0.3.3"
python-dotenv = "^1.0.1"
black = "^24.4.2"
isort = "^5.13.2"
langgraph-cli = "^0.1.35"
tiktoken = "^0.7.0"
langchain-openai = "^0.1.12"
pypdf = "^4.3.1"
faiss-cpu = "^1.8.0.post1"
langchain-chroma = "^0.1.2"
langchain = "^0.2.12"
langgraph = "^0.1.19"
langchain-community = "^0.2.11"
langchainhub = "^0.1.20"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true